fix: Update `ember-concurrency` to {dev,peer}Dependencies
Using the broad semver definition of the dependant version of
`ember-concurrency` causes problems in a yarn monorepo.

You end up with an entry in `yarn.lock` which looks something
like:

```
ember-concurrency@1.3.0, "ember-concurrency@^0.8.27 || ^0.9.0 || ^0.10.0 || ^1.0.0 || ^2.0.0-rc.1":
  version "1.3.0"
  resolved "<path/to>/ember-concurrency/-/ember-concurrency-1.3.0.tgz#66f90fb792687470bcee1172adc0ebf33f5e8b9c"
  integrity sha512-DwGlfWFpYyAkTwsedlEtK4t1DznJSculAW6Vq5S1C0shVPc5b6tTpHB2FFYisannSYkm+wpm1f1Pd40qiNPtOQ==
  dependencies:
    ember-cli-babel "^7.7.3"
    ember-compatibility-helpers "^1.2.0"
    ember-maybe-import-regenerator "^0.1.6"
```

If you try to use a combination of `ember-concurrency@v1` and
`ember-concurrency@v2` in your monorepo then things can go sideways
because you've essentially told `yarn` that they are in some way
interchangable.

I think that `devDependency` and `peerDependency` describes what
this was trying to achieve more semantically anyway (although it's
true that `peerDependencies` do _still_ suck). There are already
`ember-try` scenarios making sure it works with v1 and v2 of
`ember-concurrency`.